Added getComputerChoice and getHumanChoice functions
The getComputerChoice and getHumanChoice functions have been added as well as initializing the player and computer scores for later use.

The getComputerChoice function will auto-generate rock/paper/scissors based from a random number of 1-3. The getHumanChoice is generated by having the player enter a choice and validating that it is correct. It will also automatically change it to lowercase for easier processing.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   scripts.js